+++ benchmarks/helics/PholdFederate.hpp
@@ -5,6 +5,7 @@ the top-level NOTICE for additional details. All rights reserved.
 SPDX-License-Identifier: BSD-3-Clause
 */
 
+#include "BenchmarkFederate.hpp"
 #include "helics/application_api/Endpoints.hpp"
 #include "helics/application_api/MessageFederate.hpp"
 #include "helics/core/ActionMessage.hpp"
@@ -14,48 +15,36 @@ SPDX-License-Identifier: BSD-3-Clause
 
 using namespace helics;
 /** class implementing a federate for the PHOLD benchmark*/
-class PholdFederate {
+class PholdFederate: public BenchmarkFederate {
   public:
-    helics::Time deltaTime = helics::Time(10, time_units::ns); // sampling rate
-    helics::Time finalTime = helics::Time(10000, time_units::ns); // final time
-    int evCount = 0; // number of events handled by this federate
+    int evCount{0}; // number of events handled by this federate
 
   private:
-    std::unique_ptr<helics::MessageFederate> mFed;
-    helics::Endpoint* ept;
-
-    int index_ = 0;
-    int maxIndex_ = 0;
+    helics::Endpoint* ept{nullptr};
 
     // values from paper "Warp Speed: Executing Time Warp on 1,966,080 Cores"
     // 16 circulating events per LP (also ran with 8 events per LP)
     // 10% remote communications (also ran with 25%, 50%, and 100% remote communications)
     // mean exponential distribution of 0.90
     // optional lookahead of .1 (to prevent arbitrarily small increases in time with conservative simulators)
-    unsigned int initEvCount_ = 16; // starting number of events
-    double localProbability_ = .9; // probability of local events
-    double randTimeMean_ =
-        deltaTime * .9; // mean for the exponential distribution used when picking event times
-    double lookahead_ = deltaTime * .1;
+    unsigned int initEvCount_{16}; // starting number of events
+    double localProbability_{.9}; // probability of local events
+    double randTimeMean_{deltaTime *
+                         .9}; // mean for the exponential distribution used when picking event times
+    double lookahead_{deltaTime * .1};
 
     // classes related to the exponential and uniform distribution random number generator
-    bool generateRandomSeed = true;
-    unsigned int seed =
-        0xABad5eed; // some suggestions for seed choice were that not having a majority of the bits as 0 is better
+    bool generateRandomSeed{false};
+    unsigned int seed{
+        0xABad5eed}; // some suggestions for seed choice were that not having a majority of the bits as 0 is better
     std::mt19937 rand_gen;
     std::exponential_distribution<double> rand_exp;
     std::uniform_real_distribution<double> rand_uniform_double;
     std::uniform_int_distribution<unsigned int> rand_uniform_int;
 
-    // callbacks for instrumenting the code
-    std::function<void()> callBeforeFinalize = nullptr;
-    std::function<void()> callAfterFinalize = nullptr;
-
-    bool initialized{false};
-    bool readyToRun{false};
-
   public:
-    PholdFederate() = default;
+    // TODO: output cluster name in a summary file, along with node count and feds per node
+    PholdFederate(): BenchmarkFederate("PHOLD") {}
 
     // functions for setting parameters
     void setGenerateRandomSeed(bool b) { generateRandomSeed = b; };
@@ -63,36 +52,45 @@ class PholdFederate {
     void setRandomTimeMean(double mean) { randTimeMean_ = mean; };
     void setInitialEventCount(unsigned int count) { initEvCount_ = count; };
     void setLocalProbability(double p) { localProbability_ = p; };
+    void setLookahead(double v) { lookahead_ = v; };
 
     // functions for setting callbacks
     void setBeforeFinalizeCallback(std::function<void()> cb = nullptr) { callBeforeFinalize = cb; };
     void setAfterFinalizeCallback(std::function<void()> cb = nullptr) { callAfterFinalize = cb; };
 
-    void run(std::function<void()> callOnReady = nullptr)
-    {
-        makeReady();
-        if (callOnReady) {
-            callOnReady();
-        }
-        mainLoop();
-    };
+    std::string getName() override { return "phold_" + std::to_string(index); }
 
-    void initialize(const std::string& coreName, int index, int maxIndex)
+    void setupArgumentParsing() override
     {
-        helics::FederateInfo fi;
-        fi.coreName = coreName;
-        initialize(fi, index, maxIndex);
+        app->add_option("--init_ev_count", initEvCount_, "the starting number of events");
+        app->add_option(
+            "--local_probability", localProbability_, "the probability of local events");
+        app->add_option(
+            "--rand_time_mean",
+            randTimeMean_,
+            "mean for the exponential distribution used when picking event times");
+        app->add_flag("--gen_rand_seed", generateRandomSeed, "enable generating a random seed");
+        app->add_option("--set_rand_seed", seed, "set the random seed");
+        app->add_option("--set_phold_lookahead", lookahead_, "set the lookahead used by phold");
     }
 
-    void initialize(const helics::FederateInfo fi, int index, int maxIndex)
+    void doAddBenchmarkResults() override
     {
-        std::string name = "phold_" + std::to_string(index);
-        index_ = index;
-        maxIndex_ = maxIndex;
+        addResult("EVENT COUNT", "EvCount", std::to_string(evCount));
+    }
 
-        mFed = std::make_unique<helics::MessageFederate>(name, fi);
-        ept = &mFed->registerEndpoint("ept");
+    void doParamInit(helics::FederateInfo& /*fi*/) override
+    {
+        if (app->get_option("--set_rand_seed")->count() == 0) {
+            std::mt19937 rand_gen(0x600d5eed);
+            std::uniform_int_distribution<unsigned int> rand_seed_uniform;
+            for (int i = 0; i < index; i++) {
+                rand_seed_uniform(rand_gen);
+            }
+            setRandomSeed(rand_seed_uniform(rand_gen));
+        }
 
+        // set up based on given params
         // en.cppreference.com/w/cpp/numeric/random/exponential_distribution
         if (generateRandomSeed) {
             std::random_device rd;
@@ -103,63 +101,27 @@ class PholdFederate {
         rand_exp = std::exponential_distribution<double>(1.0 / randTimeMean_);
         rand_uniform_double = std::uniform_real_distribution<double>(0.0, 1.0);
         // create random number distribution for picking a destination if there is more than 1 federate
-        if (maxIndex_ > 1) {
-            rand_uniform_int = std::uniform_int_distribution<unsigned int>(0, maxIndex_ - 2);
-        }
-        initialized = true;
-    }
-
-    void finalize()
-    {
-        if (callBeforeFinalize) {
-            callBeforeFinalize();
-        }
-        mFed->finalize();
-        if (callAfterFinalize) {
-            callAfterFinalize();
+        if (maxIndex > 1) {
+            rand_uniform_int = std::uniform_int_distribution<unsigned int>(0, maxIndex - 2);
         }
     }
 
-    void makeReady()
-    {
-        if (!initialized) {
-            throw("must initialize first");
-        }
-        mFed->enterExecutingMode();
+    void doFedInit() override { ept = &fed->registerEndpoint("ept"); }
 
+    void doMakeReady() override
+    {
         // create initial events (not included in benchmark time)
         for (unsigned int i = 0; i < initEvCount_; i++) {
             createNewEvent();
         }
-
-        readyToRun = true;
     }
 
-    void createNewEvent()
-    {
-        // decide if the event is local or remote
-        auto destIndex = index_;
-        if (maxIndex_ > 1 && rand_uniform_double(rand_gen) > localProbability_) {
-            destIndex = rand_uniform_int(rand_gen);
-            if (destIndex == index_) {
-                destIndex = maxIndex_ - 1;
-            }
-        }
-
-        // set the event time to current time + lookahead + rand exponential (mean >= lookahead or ~2x lookahead)
-        helics::Time evTime =
-            mFed->getCurrentTime() + helics::Time(lookahead_) + helics::Time(rand_exp(rand_gen));
-        std::string data = "ev";
-        std::string dest = "phold_" + std::to_string(destIndex) + "/ept";
-        ept->send(dest, data, evTime);
-    }
-
-    void mainLoop()
+    void doMainLoop() override
     {
         auto nextTime = deltaTime;
 
         while (nextTime < finalTime) {
-            nextTime = mFed->requestTime(finalTime);
+            nextTime = fed->requestTime(finalTime);
             // for each event message received, create a new event
             while (ept->hasMessage()) {
                 auto m = ept->getMessage();
@@ -167,6 +129,24 @@ class PholdFederate {
                 createNewEvent();
             }
         }
-        finalize();
+    }
+
+    void createNewEvent()
+    {
+        // decide if the event is local or remote
+        auto destIndex = index;
+        if (maxIndex > 1 && rand_uniform_double(rand_gen) > localProbability_) {
+            destIndex = rand_uniform_int(rand_gen);
+            if (destIndex == index) {
+                destIndex = maxIndex - 1;
+            }
+        }
+
+        // set the event time to current time + lookahead + rand exponential (mean >= lookahead or ~2x lookahead)
+        helics::Time evTime =
+            fed->getCurrentTime() + helics::Time(lookahead_) + helics::Time(rand_exp(rand_gen));
+        std::string data = "ev";
+        std::string dest = "phold_" + std::to_string(destIndex) + "/ept";
+        ept->send(dest, data, evTime);
     }
 };