+++ benchmarks/helics/echoMessageBenchmarks.cpp
@@ -5,7 +5,8 @@ the top-level NOTICE for additional details. All rights reserved.
 SPDX-License-Identifier: BSD-3-Clause
 */
 
-#include "EchoMessage.hpp"
+#include "EchoMessageHubFederate.hpp"
+#include "EchoMessageLeafFederate.hpp"
 #include "helics/core/BrokerFactory.hpp"
 #include "helics/core/CoreFactory.hpp"
 #include "helics/helics-config.h"
@@ -31,10 +32,11 @@ static void BMecho_singleCore(benchmark::State& state)
         auto wcore = helics::CoreFactory::create(
             core_type::INPROC, std::string("--autobroker --federates=") + std::to_string(feds + 1));
         EchoMessageHub hub;
-        hub.initialize(wcore->getIdentifier());
+        hub.initialize(wcore->getIdentifier(), "");
         std::vector<EchoMessageLeaf> leafs(feds);
         for (int ii = 0; ii < feds; ++ii) {
-            leafs[ii].initialize(wcore->getIdentifier(), ii);
+            std::string bmInit = "--index=" + std::to_string(ii);
+            leafs[ii].initialize(wcore->getIdentifier(), bmInit);
         }
 
         std::vector<std::thread> threadlist(static_cast<size_t>(feds));
@@ -79,13 +81,14 @@ static void BMecho_multiCore(benchmark::State& state, core_type cType)
             helics::CoreFactory::create(cType, std::string("--federates=1 --log_level=no_print"));
         // this is to delay until the threads are ready
         EchoMessageHub hub;
-        hub.initialize(wcore->getIdentifier());
+        hub.initialize(wcore->getIdentifier(), "");
         std::vector<EchoMessageLeaf> leafs(feds);
         std::vector<std::shared_ptr<helics::Core>> cores(feds);
         for (int ii = 0; ii < feds; ++ii) {
             cores[ii] = helics::CoreFactory::create(cType, "-f 1 --log_level=no_print");
             cores[ii]->connect();
-            leafs[ii].initialize(cores[ii]->getIdentifier(), ii);
+            std::string bmInit = "--index=" + std::to_string(ii);
+            leafs[ii].initialize(cores[ii]->getIdentifier(), bmInit);
         }
 
         std::vector<std::thread> threadlist(static_cast<size_t>(feds));